move panel view to ui app
