+++ ikarodjango/ui/urls.py
@@ -1,7 +1,7 @@
 from django.urls import path
 
-from ui.views.pages import Home
-from panel.views import FlightPanel
+from ui.views.pages import Home, FlightPanel
+
 
 urlpatterns = [
     path('', Home.as_view(), name='home'),
